+++ packages/flame/test/extensions/canvas_test.dart
@@ -8,8 +8,8 @@ void main() {
       final canvas = MockCanvas();
       canvas.renderPoint(Vector2.all(10.0), size: 2);
       expect(
-        canvas.methodCalls,
-        contains('drawRect(9.0, 9.0, 2.0, 2.0)'),
+        canvas,
+        MockCanvas()..drawRect(const Rect.fromLTWH(9, 9, 2, 2)),
       );
     });
   });