+++ packages/flame/CHANGELOG.md
@@ -40,6 +40,7 @@
  - In debug mode `PositionComponent` now displays an indicator for the anchor position.
  - Update `Camera` docs to showcase usage with `Game` class
  - Fixed a bug with `worldBounds` being set to `null` in `Camera`
+ - `MockCanvas` is now strongly typed and matches numeric coordinates up to a tolerance
 
 ## [1.0.0-releasecandidate.13]
  - Fix camera not ending up in the correct position on long jumps